refactor(docker): prepare docker for lib packaging
