+++ app/scripts/app.ts
@@ -1,11 +1,13 @@
-import * as angular from 'angular';
+import { module } from 'angular';
 
-import {NETFLIX_MODULE} from './modules/netflix/netflix.module';
-import {APPENGINE_MODULE} from './modules/appengine/appengine.module';
+import { CORE_MODULE } from '@spinnaker/core';
+import { DOCKER_MODULE } from '@spinnaker/docker';
+import { APPENGINE_MODULE } from './modules/appengine/appengine.module';
+import { NETFLIX_MODULE } from './modules/netflix/netflix.module';
 
-module.exports = angular.module('netflix.spinnaker', [
+module('netflix.spinnaker', [
   NETFLIX_MODULE,
-  require('./modules/core/src/core.module.js'),
+  CORE_MODULE,
   require('./modules/amazon/aws.module.js'),
   require('./modules/google/gce.module.js'),
   require('./modules/cloudfoundry/cf.module.js'),
@@ -13,7 +15,7 @@ module.exports = angular.module('netflix.spinnaker', [
   require('./modules/azure/azure.module.js'),
   require('./modules/kubernetes/kubernetes.module.js'),
   require('./modules/openstack/openstack.module.js'),
-  require('./modules/docker/docker.module.js'),
+  DOCKER_MODULE,
   require('./modules/oracle/oraclebmcs.module.js'),
   APPENGINE_MODULE,
 ]);
