+++ app/scripts/modules/core/src/index.ts
@@ -1,3 +1,5 @@
+export const CORE_MODULE = 'spinnaker.core';
+
 export * from './account';
 export * from './api';
 export * from './application';
@@ -55,3 +57,5 @@ export * from './task';
 export * from './utils';
 
 export * from './widgets';
+
+require('./core.module');
