--- app/scripts/modules/docker/docker.module.js
@@ -1,14 +1,0 @@
-'use strict';
-
-const angular = require('angular');
-
-// load all templates into the $templateCache
-let templates = require.context('./', true, /\.html$/);
-templates.keys().forEach(function(key) {
-  templates(key);
-});
-
-module.exports = angular.module('spinnaker.deck.docker', [
-  require('./pipeline/stages/bake/dockerBakeStage.js'),
-  require('./pipeline/trigger/dockerTrigger.module'),
-]);
