+++ app/scripts/modules/docker/src/image/docker.image.reader.service.ts
