+++ app/scripts/modules/docker/src/image/dockerImageAndTagSelector.component.html
