+++ app/scripts/modules/docker/src/image/dockerImageAndTagSelector.component.spec.ts
