+++ app/scripts/modules/docker/src/image/index.ts
@@ -1,0 +1,1 @@
+export * from './docker.image.reader.service';
