+++ app/scripts/modules/docker/src/index.ts
@@ -1,0 +1,2 @@
+export * from './image';
+export * from './docker.module';
