+++ app/scripts/modules/docker/src/pipeline/stages/bake/bakeExecutionDetails.controller.js
