+++ app/scripts/modules/docker/src/pipeline/stages/bake/bakeStage.html
