+++ app/scripts/modules/docker/src/pipeline/stages/bake/dockerBakeStage.js
