+++ app/scripts/modules/docker/src/pipeline/trigger/dockerTrigger.html
