+++ app/scripts/modules/docker/src/pipeline/trigger/dockerTrigger.module.js
