+++ app/scripts/modules/docker/src/pipeline/trigger/dockerTriggerOptions.directive.html
