+++ app/scripts/modules/docker/src/pipeline/trigger/selectorTemplate.html
