+++ app/scripts/modules/kubernetes/pipeline/stages/runJob/runJobStage.js
@@ -4,10 +4,7 @@ import _ from 'lodash';
 
 const angular = require('angular');
 
-import {DOCKER_IMAGE_AND_TAG_SELECTOR_COMPONENT} from 'docker/image/dockerImageAndTagSelector.component';
-
 module.exports = angular.module('spinnaker.kubernetes.pipeline.stage.runJobStage', [
-  DOCKER_IMAGE_AND_TAG_SELECTOR_COMPONENT,
   require('kubernetes/container/commands.component.js'),
   require('kubernetes/container/arguments.component.js'),
   require('kubernetes/container/environmentVariables.component.js'),
